Include leave in stats
