+++ app/templates/pages/home.html.j2
@@ -3,7 +3,7 @@
 {% set date_fmt = 'Do [of] MMM YYYY [at] HH:mm' %}
 
     <!-- Header Stats -->
-    <dynamic-frame id="time-stats" :execute-scripts="true" :url="/frames/time-stats">
+    <dynamic-frame id="stats" :execute-scripts="true" :url="/frames/stats">
     </dynamic-frame>
 
     <!-- Clock in/out Form -->
